refactor: move runnable apps under apps/ and flatten the library source
Everything runnable now sits in apps/: the playground joins the docs site.
Inside src/, the twelve components collapse to one level, GanttProps becomes
types/props.ts, the stylesheet moves out of a three-deep assets/ directory,
and the barrel loses its .tsx extension.

Also fixes a packaging bug: three emitted .d.ts files imported from a bare
'core', which no consumer could resolve. A CI step now guards against it.
