+++ apps/docs/content/docs/ref/core-critical-path.en.mdx
@@ -199,7 +199,7 @@ With no `projectFinish` argument and an empty `early` map, the function returns 
 
 ## Not exported from the package root
 
-These names appear above but are **not** in `src/index.tsx`, so they cannot be imported from
+These names appear above but are **not** in `src/index.ts`, so they cannot be imported from
 `@jaeungkim/gantt-chart`:
 
 | Name | Kind | Workaround |
