+++ apps/docs/content/docs/ref/core-critical-path.ko.mdx
@@ -197,7 +197,7 @@ export interface LateDates {
 
 ## 패키지 루트에서 내보내지 않는 것
 
-위에 나온 이름들 중 이것들은 `src/index.tsx`에 **없어요**. 그래서 `@jaeungkim/gantt-chart`에서
+위에 나온 이름들 중 이것들은 `src/index.ts`에 **없어요**. 그래서 `@jaeungkim/gantt-chart`에서
 import할 수 없어요.
 
 | 이름 | 종류 | 우회 방법 |
