+++ apps/docs/content/docs/ref/interaction-config.en.mdx
@@ -217,7 +217,7 @@ export function canCreateTasks(
 with the presence of the callback:
 
 ```ts
-// src/components/Gantt/Gantt.tsx
+// src/components/Gantt.tsx
 const canDrawTasks = onTaskCreate !== undefined && canCreateTasks(interaction);
 ```
