+++ apps/docs/content/docs/ref/interaction-config.ko.mdx
@@ -214,7 +214,7 @@ export function canCreateTasks(
 AND로 묶어요:
 
 ```ts
-// src/components/Gantt/Gantt.tsx
+// src/components/Gantt.tsx
 const canDrawTasks = onTaskCreate !== undefined && canCreateTasks(interaction);
 ```
