+++ apps/docs/content/docs/ref/markers.en.mdx
@@ -114,7 +114,7 @@ A task ending exactly on `date` is not an overrun.
 The chart prepends one marker of its own to whatever `markers` holds:
 
 ```ts
-// src/components/Gantt/Gantt.tsx
+// src/components/Gantt.tsx
 { id: "today", date: dayjs(), className: "gantt-today-marker" }
 ```
