+++ pnpm-workspace.yaml
@@ -1,6 +1,5 @@
 packages:
   - apps/*
-  - playground
 
 # esbuild ships prebuilt binaries via optionalDependencies; its postinstall script is not needed.
 allowBuilds:
