+++ src/components/GanttBarsLayer.tsx
