+++ src/components/GanttNonWorkingLayer.tsx
